pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "repro-llm-stack"
version = "0.1.0"
description = "Reproducible ~7B LM stack: data, train, eval, serve"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "You" }]
dependencies = [
  "pyyaml>=6.0.1",
]

[project.optional-dependencies]
dev = [
  "ruff>=0.6.0",
]
data = [
  "datasets>=2.19.0",
  "huggingface_hub>=0.23.0",
]
train = [
  "datasets>=2.19.0",
  "huggingface_hub>=0.23.0",
  "llamafactory>=0.9.0",
  "torch>=2.3.0",
  "transformers>=4.44.0",
  "accelerate>=0.33.0",
]
eval = [
  "lm-eval>=0.4.4",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["tools*"]

[tool.ruff]
line-length = 100
target-version = "py310"
